Add files via upload
Rewrote the method for generating the chunk mesh and texturing it. Now using a dumb greedy meshing algotithm and texture arrays. Should notice performance increase. Temporary removed physics and saving to file is not working. Big bug with Frustum Culling: code that was written but later removed is being executed even if it's not in the files.